third_party/blink/web_tests/fast/css/forced-colors-mode/forced-colors-mode-11-expected.txt

This is a testharness.js-based test.
FAIL Checks that styles defined in forced-colors are preserved in forced colors mode. assert_equals: expected "rgb(0, 0, 255)" but got "rgba(0, 0, 0, 0)"
Harness: the test ran to completion.
